[CL-2084] Move components and containers from project folder module to main app
