+++ front/app/components/ProjectFolders/ProjectFolderSharingModal/messages.ts
