+++ front/app/containers/ProjectFolderShowPage/ProjectFolderHeader.tsx
@@ -2,7 +2,7 @@ import React, { memo } from 'react';
 
 // components
 import Image from 'components/UI/Image';
-import ProjectFolderShareButton from '../../../citizen/components/ProjectFolderShareButton';
+import ProjectFolderShareButton from '../../components/ProjectFolders/ProjectFolderShareButton';
 
 // hooks
 import { useWindowSize } from '@citizenlab/cl2-component-library';
