+++ front/app/containers/ProjectFolderShowPage/styles.ts
